# apps/api/openapi/v1.yaml
...
openapi: "3.1.0"
info:
  title: EasyA2 API — Foundation + Org Structure + Identity Advanced + Process Modeling
  version: "1.4.0"
  description: |
    Foundation (MOD-000) + Organizational Structure (MOD-003) + Identity Advanced (MOD-004) + Process Modeling (MOD-005) API.
    All errors follow RFC 9457 Problem Details.
servers:
  - url: /api/v1
...
paths:
...
  /users:
    post:
      operationId: users_create
      tags: [users]
      summary: Register new user (public or admin)
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateUserResponse"
        "422":
          $ref: "#/components/responses/ProblemDetails422"
    get:
      operationId: users_list
      tags: [users]
      summary: List users (paginated)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
  /users/{id}:
    get:
      operationId: users_get
      tags: [users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
        "404":
          $ref: "#/components/responses/ProblemDetails404"
    patch:
      operationId: users_update
      tags: [users]
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
    delete:
      operationId: users_delete
      tags: [users]
      summary: Soft-delete user (LGPD)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: User soft-deleted

  # ---------- Roles ---------------------------------------------------------
...
  /tenants/{tenantId}/users:
    post:
      operationId: tenant_users_add
      tags: [tenant-users]
      summary: Link user to tenant with role
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantIdPath"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddTenantUserRequest"
      responses:
        "201":
          description: User linked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantUserAddResponse"
    get:
      operationId: tenant_users_list
      tags: [tenant-users]
      summary: List users of a tenant
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantIdPath"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Tenant user list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTenantUsers"
  /tenants/{tenantId}/users/{userId}:
    delete:
      operationId: tenant_users_remove
      tags: [tenant-users]
      summary: Unlink user from tenant
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantIdPath"
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: User unlinked

  # ---------- System --------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TenantIdPath:
      name: tenantId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema:
        type: string
        format: uuid
      description: Idempotency key to prevent duplicate mutations (BR-012)
...
  responses:
...
    ProblemDetails404:
      description: Not Found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
    ProblemDetails422:
      description: Unprocessable Entity
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
  schemas:
    ProblemDetails:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        extensions:
          type: object
          properties:
            correlationId: { type: string }
...
    CreateUserRequest:
      type: object
      required: [email, password, full_name]
      properties:
        email: { type: string, format: email, maxLength: 255 }
        password: { type: string, minLength: 8, maxLength: 128 }
        full_name: { type: string, minLength: 1, maxLength: 255 }
        cpf_cnpj: { type: string, maxLength: 20 }
    CreateUserResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        codigo: { type: string }
        full_name: { type: string }
        status: { type: string }
    PaginatedUsers:
      type: object
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/UserListItem" } }
        next_cursor: { type: string, nullable: true }
        has_more: { type: boolean }
    UserListItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        codigo: { type: string }
        email: { type: string }
        full_name: { type: string }
        status: { type: string }
        created_at: { type: string, format: date-time }
    UserDetail:
      type: object
      properties:
        id: { type: string, format: uuid }
        codigo: { type: string }
        email: { type: string }
        full_name: { type: string }
        cpf_cnpj: { type: string, nullable: true }
        avatar_url: { type: string, nullable: true }
        status: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    UpdateUserRequest:
      type: object
      properties:
        full_name: { type: string }
        cpf_cnpj: { type: string, nullable: true }
        status: { type: string, enum: [ACTIVE, BLOCKED, PENDING, INACTIVE] }
...
    AddTenantUserRequest:
      type: object
      required: [user_id, role_id]
      properties:
        user_id: { type: string, format: uuid }
        role_id: { type: string, format: uuid }

    # ---- Response schemas (PENDENTE-016) ----------------------------------------
...
    TenantUserAddResponse:
      type: object
      properties:
        user_id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        role_id: { type: string, format: uuid }
        status: { type: string }
        created_at: { type: string, format: date-time }
    TenantUserListItem:
      type: object
      properties:
        user_id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        role_id: { type: string, format: uuid }
        role_name: { type: string }
        user_email: { type: string }
        user_full_name: { type: string }
        status: { type: string }
        created_at: { type: string, format: date-time }
    PaginatedTenantUsers:
      type: object
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/TenantUserListItem" } }
        next_cursor: { type: string, nullable: true }
        has_more: { type: boolean }
...